Update and rename The Battle of Neightborhoods Week 1. COVID-19 Costa Rica to The Battle of Neightborhoods Week 1. Finding Nice locations to open a restaurant
1. Introduction ¶
1.1 Background
Costa Rica is well known country where you can find people from all around the world. With COVID-19 pandemic, many restaurants and different businesses closed due to the lack of money in the streets and people losing their jobs. Since we have many tourists coming in the next months trying to reactivate the economic activity, this is a good option for our country to find a new stability for those families that need new and fast incomes. 
1.2 Problem
Many families need to support themselves and need immediate incomes with a low risk business as fast as possible, making sure the money they have left is going to support their families again. A restaurant is the kind of business that may work for them, everybody needs to eat and tourism is the main activity in Costa Rica, so we are going to find nice locations where they can risk their money and make sure their families are going to be OK.